requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
ansicon==1.89.0
arrow==1.3.0
asgiref==3.7.2
attrs==23.2.0
blessed==1.20.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2
django-allauth==0.57.0
django-compat==1.0.15
django-cors-headers==4.3.1
django-picklefield==3.1
djangorestframework==3.14.0
frozenlist==1.4.1
idna==3.6
jinxed==1.2.1
multidict==6.0.4
numpy==1.26.3
oauthlib==3.2.2
pandas==2.2.0
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
redis==3.5.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
twilio==8.11.1
types-python-dateutil==2.8.19.20240106
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
xlrd==2.0.1
yarl==1.9.4
